requirements.txt
anyio==3.6.2
attrs==22.2.0
beautifulsoup4==4.11.2
certifi==2022.12.7
charset-normalizer==3.0.1
colorama==0.4.6
exceptiongroup==1.1.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
iniconfig==2.0.0
numpy==1.24.2
packaging==23.0
pandas==1.5.3
pluggy==1.0.0
pypdf==3.4.1
pytest==7.2.1
python-dateutil==2.8.2
python-decouple==3.7
pytz==2022.7.1
requests==2.28.2
responses==0.22.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.4
toml==0.10.2
tomli==2.0.1
types-toml==0.10.8.4
typing_extensions==4.5.0
urllib3==1.26.14
